Исправила генерацию немного
